third_party/WebKit/LayoutTests/webaudio/realtimeanalyser-downmix-expected.txt

Test AnalyserNode Downmixing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Analyser downmix mono to mono time data is identical to the array [1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,...].
PASS Analyser downmix mono to mono freq data equals [-7.53501,-12.0412,-27.9588,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 6.3283e-8}.
PASS Analyser downmix mono to mono downmixed correctly.

PASS Analyser downmix stereo to mono time data is identical to the array [2.5,2.5,2.5,2.5,2.5,2.5,2.5,2.5,2.5,2.5,2.5,2.5,2.5,2.5,2.5,2.5,...].
PASS Analyser downmix stereo to mono freq data equals [0.423786,-4.08240,-20.0000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 1.1681e-7}.
PASS Analyser downmix stereo to mono downmixed correctly.

PASS Analyser downmix quad to mono time data is identical to the array [7.5,7.5,7.5,7.5,7.5,7.5,7.5,7.5,7.5,7.5,7.5,7.5,7.5,7.5,7.5,7.5,...].
PASS Analyser downmix quad to mono freq data equals [9.96621,5.46002,-10.4576,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 4.9793e-7}.
PASS Analyser downmix quad to mono downmixed correctly.

PASS Analyser downmix 5.1 to mono time data is identical to the array [43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,43.0355339050293,...].
PASS Analyser downmix 5.1 to mono freq data equals [25.1415,20.6353,4.71774,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 2.0215e-7}.
PASS Analyser downmix 5.1 to mono downmixed correctly.

PASS Analyser downmix 3-channel to mono time data is identical to the array [1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,...].
PASS Analyser downmix 3-channel to mono freq data equals [-7.53501,-12.0412,-27.9588,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,-100.000,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 6.3283e-8}.
PASS Analyser downmix 3-channel to mono downmixed correctly.

PASS successfullyParsed is true

TEST COMPLETE
